working. update int types.

+++ include/data.h
@@ -1,17 +1,19 @@
 #include "config.h"
 #include "encode.h"
+#include <stddef.h>
+#include <stdint.h>
 
 #ifndef DATA_H
 #define DATA_H
 
 typedef struct {
   float *membranes;
-  int *conns;
+  size_t *conns;
   float *weights;
   float *thresholds;
-  int *post_spikes;
-  int *pre_spikes;
-  int *refactory;
+  uint8_t *post_spikes;
+  uint8_t *pre_spikes;
+  uint8_t *refactory;
   float *pre_trace;
   float *post_trace;
 } NetworkData;
