+++ include/utils.h
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
-int *generate_connections(int in_n, int out_n, int n_conns);
+size_t *generate_connections(int in_n, int out_n, int n_conns);
 // int *generate_connections(int n_neurons, int n_conns);
 
 float *generate_weights(int in_n, int n_conns, float min, float max);
